+++ CHANGELOG.md
@@ -507,3 +507,4 @@
 - 2026-04-21: Test summaries, mod, send, site config, +2
 - 2026-04-22: Test embeddings, mod, summaries, send, +1
 - 2026-04-22: Integ summaries, mod, embeddings, send, +1
+- 2026-04-23: Cache mod, mod, summaries, embeddings, +1